Adding provider-specific request timeout
the request timeout which is used for both the timeout of a
single request and the retry-timeout-limit, can now be set in
the provider config view for each provider.

The input accepts integer values between 1 & 3600 and defaults
to 60 seconds for new providers or in case of no input.
Values outside the range are automatically corrected to the
bottom or top value.

- add timeout field & validation to provider config
- client.py reads timeout from provider config
- propagate timeout to NetworkAccessManager
- adjust default config
- add CHANGELOG entry

refers to https://github.com/GIScience/orstools-qgis-plugin/issues/122

Co-authored-by: Jakob Schnell <jakob.schnell@heigit.org>
Co-authored-by: Amandus Butzer <amandus.butzer@heigit.org>