Get and set active provider in processing algorithms
to use the active provider in processing algorithms by default,
it is read from the storage.
However, the defaultValue for QgsProcessingParameterEnum is not working,
which is why the options of the enum are reordered to show the active
provider on top.
As the index of the provider selected in the Enum does not refer to the
same provider in the config anymore, it has to be resolved properly again,
when using the provider to generate the ors_client.

This can lead to problems (using the wrong provider) when the active provider
is changed, after a processing algorithm window was opened, due to how
the algorithm parameters are initialized.